Change tag key to only use 'level'
Tags in Humio is almost a hidden feature and should be used with
care. They create files in the order of the Cartesian product in tag
values, so that if you by accident puts some kind of id (pid, request
id, ...) into a tag, things will explode making the Humio data space
unusable. This is a severe limitation in the current tagging scheme in
Humio and we are working on a better way to tag/shard data in the near
future. Until then, this commit will ensures that we avoid a tag
explosion, by only tagging on the log-level. Host and source
information has been moved to the regular attributes.